update config for 1.21

+++ src/main/java/fi/dy/masa/minihud/data/EntitiesDataManager.java
@@ -213,7 +213,7 @@ public class EntitiesDataManager implements IClientTickHandler
 
     private long getCacheTimeout()
     {
-        return (long) (MathHelper.clamp(Configs.Generic.ENTITY_DATA_SYNC_CACHE_TIMEOUT.getFloatValue(), 0.25f, 25.0f) * 1000L);
+        return (long) (MathHelper.clamp(Configs.Generic.ENTITY_DATA_SYNC_CACHE_TIMEOUT.getDoubleValue(), 0.25f, 25.0f) * 1000L);
     }
 
     private void tickCache()
